+++ apps/web/src/app-routes.tsx
@@ -5,10 +5,12 @@ import { DashboardPage1 } from './features/dashboard/pages/dashboard-page1'
 import { DashboardPage2 } from './features/dashboard/pages/dashboard-page2'
 
 // routes
-export const LandingPageRouteId = app.router.bind({ path: '/', element: <LandingPage /> })
-export const DashboardRouteId = app.router.bind({ path: '/dashboard', element: <DashboardLayout /> })
-export const DashboardPage1RouteId = app.router.bind(
-    { path: 'page1', index: true, Component: DashboardPage1 },
-    DashboardRouteId
-)
-export const DashboardPage2RouteId = app.router.bind({ path: 'page2', Component: DashboardPage2 }, DashboardRouteId)
+export const LandingPageRouteId = app.routes.add({ path: '/', element: <LandingPage /> })
+
+export const DashboardRouteId = app.routes.add({ path: '/dashboard', element: <DashboardLayout /> })
+export const DashboardPage1RouteId = DashboardRouteId.addChild({
+    path: 'page1',
+    index: true,
+    Component: DashboardPage1
+})
+export const DashboardPage2RouteId = DashboardRouteId.addChild({ path: 'page2', Component: DashboardPage2 })