+++ apps/web/src/app-stores.tsx
@@ -2,6 +2,6 @@ import { app } from './app'
 import { CounterStore } from './features/dashboard/store/counter.store'
 
 // stores
-export const ICounterStore = app.stores.bind(CounterStore)
+export const ICounterStore = app.stores.add(CounterStore)
 
 // services