+++ apps/web/src/main.tsx
@@ -14,6 +14,7 @@ const SuspenseFallbackComponent = () => <div></div>
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
     <React.StrictMode>
+        {/* Other Custom Providers Wrap App */}
         <VinessReactApp
             app={app}
             SuspenseFallbackComponent={SuspenseFallbackComponent}