+++ packages/viness-react/src/app-react-hooks.tsx
@@ -1,9 +1,6 @@
-import { ServiceInstanceIdentifier } from '@viness/di'
-import { VinessApp } from './app'
+import { VinessApp } from './app/app'
 import { useAppContext } from './app-react-context'
 import { VinessServiceIdentifier } from './identifier'
-import { IVinessRoute } from './route'
-import { UIStore } from './store'
 
 /**
  * Get the Viness App instance
@@ -19,51 +16,4 @@ export function useApp() {
 export function useService<T>(id: VinessServiceIdentifier<T>) {
     const app = useAppContext()
     return app?.services.get(id) as T
-}
-
-/**
- * Get 'Router' instance in FC
- *
- * @returns {IVinessRouter}
- */
-export const useRouter = () => {
-    const app = useAppContext()
-    return app.router
-}
-
-/**
- * Get specific Route Instance by id in FC
- *
- * @param id
- * @returns {IVinessRoute}
- */
-export function useRoute(id: VinessServiceIdentifier<IVinessRoute>) {
-    const app = useAppContext()
-    return app?.router?.get(id)
-}
-
-/**
- * Get a list of route instances by route ids passed
- *
- * @param ids
- * @returns
- */
-export function useRoutes(ids: VinessServiceIdentifier<IVinessRoute>[]) {
-    const app = useAppContext()
-    const routes = ids.map((id) => app?.router?.get(id))
-    return routes
-}
-
-/**
- * Get UIStore instance
- *
- * @param {ServiceIdentifier<T>} id
- * @returns {T}
- */
-export function useStore<S extends object, T extends UIStore<S>>(
-    id: VinessServiceIdentifier<T>,
-    instanceId?: ServiceInstanceIdentifier
-) {
-    const app = useAppContext()
-    return app?.stores.get(id, instanceId)
 }