+++ packages/viness-react/src/app/app-config.tsx
@@ -1,6 +1,6 @@
-import { VinessI18nConfig } from './i18n'
-import { createIdentifier } from './identifier'
-import { VinessRouterConfig } from './router'
+import { VinessI18nConfig } from '../i18n/i18n'
+import { createIdentifier } from '../identifier'
+import { VinessRouterConfig } from '../route/routes'
 
 export const IVinessAppConfig = createIdentifier<VinessAppConfig>('IVinessAppConfig')
 
@@ -13,7 +13,7 @@ export class VinessAppConfig implements IVinessAppConfig {
     readonly i18n?: VinessI18nConfig
     readonly router?: VinessRouterConfig
 
-    constructor(config: IVinessAppConfig) {
+    constructor(config?: IVinessAppConfig) {
         if (!config) return
         const { i18n, router } = config
         this.i18n = i18n