+++ packages/viness-react/src/container.ts
@@ -2,4 +2,3 @@ import { Container } from '@viness/di'
 
 export const storeContainer = new Container()
 export const servicesContainer = storeContainer.createChild()
-export const effectsContainer = servicesContainer.createChild()