+++ packages/viness-react/src/http/http-client.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosRequestConfig, AxiosInstance } from 'axios'
-import { createIdentifier } from './identifier'
+import { createIdentifier } from '../identifier'
 
 /**
  * Promise can be canceled